phylum-cli: adhere to formula standards
This change ensures the Phylum CLI is not built with the ability to
upgrade or uninstall itself, leaving those functions to `brew` instead.
It also generates and installs shell completion files. A caveat was
added to inform users that official extensions were not preinstalled.
Finally, the version command used for testing was altered to the form
that provides more succinct output so as not to overmatch on the assert.