cointop: deprecate
The GitHub repository for `cointop` was archived on 2023-11-18. The
most recent release (v1.6.10) was on 2021-11-06 and the most recent
commit was on 2022-08-31. The homepage (https://cointop.sh) isn't
reachable at this point.

The `README` wasn't updated to explain the project status before the
repository was archived but this seems to suggest that it isn't being
developed/maintained further. This deprecates the formula accordingly.