dvd+rw-tools: update urls, license
The `homepage`, `stable`, and `livecheck` URLs for `dvd+rw-tools`
redirect from HTTP to HTTPS. This updates the URLs to avoid the
redirection.

This also updates the license from `GPL-2.0` to `GPL-2.0-only`. The
only usage of the "or...later" language in the project files is some
Hewlett-Packard code in `transport.hxx`, so this seems like
`GPL-2.0-only` by default.