git-test: deprecate
The GitHub repository for `git-test`
(https://github.com/spotify/git-test) was archived on 2023-01-11. The
most recent tag (v1.0.4) was on 2016-08-17 and the most recent commit
was on 2020-08-07. The `README` wasn't updated to explain the project
status before the repository was archived but this presumably means
that the project isn't being developed/maintained further. This
deprecates the formula accordingly.