hubflow: deprecate, update hompage
The GitHub repository for `hubflow`
https://github.com/datasift/gitflow was archived on 2023-06-19. The
most recent tag (1.5.4) was on 2022-04-29 and the most recent commit
was on 2022-05-01. The `README` wasn't updated to explain the project
status before the repository was archived but this presumably means
that the project isn't being developed/maintained further. This
deprecates the formula accordingly.

This also updates the homepage to the GitHub repository, as the
github.io homepage currently returns a 404 (Not Found) response.