launchdns: deprecate
The GitHub repository for `launchdns`
(https://github.com/josh/launchdns) was archived on 2023-05-22. The
most recent tag (v1.0.4) and commit was on 2019-08-23. The `README`
wasn't updated to explain the project status before the repository
was archived but this presumably means that the project isn't being
developed/maintained further. This deprecates the formula accordingly.