msgpuck: deprecate
The GitHub repository for `msgpuck`
(https://github.com/rtsisyk/msgpuck) was archived on 2023-03-01. The
most recent tag (2.0) was on 2017-02-07 and the most recent commit
was on 2019-03-12. The `README` wasn't updated to explain the project
status before the repository was archived but this presumably means
that the project isn't being developed/maintained further. This
deprecates the formula accordingly.