opentracing-cpp: deprecate
The GitHub repository for `opentracing-cpp`
(https://github.com/opentracing/opentracing-cpp) was archived on
2024-01-26. The most recent release (1.6.0) was on 2019-09-19 and the
most recent commit prior to deprecation was on 2021-07-21. Before the
repository was archived, the `README` was updated to add:

> 🛑 This library is DEPRECATED!
> https://github.com/opentracing/specification/issues/163

The linked issue from 2022-01-14 proposed that development should
cease, as practically all activity moved to the OpenTelemetry project.

Since the project isn't being developed/maintained further, this
deprecates the formula accordingly.