pbzx: deprecate
The GitHub repository for `pbzx`
(https://github.com/NiklasRosenstein/pbzx) was archived on
2023-11-06. The most recent release (v1.0.2) was on 2017-01-12 and
the most recent [non-documentation] commit was on 2017-01-11. The
`README` wasn't updated to explain the project status before the
repository was archived but this presumably means that the project
isn't being developed/maintained further. This deprecates the
formula accordingly.