tm: deprecate
The GitHub repository for `tm` (https://github.com/triggermesh/tm)
was archived on 2023-05-09. The most recent tag (v1.21.0) was on
2022-09-05 and the most recent commit prior to editing the `README`
was on 2022-07-11. The `README` was updated before archiving with
the following message:

> THIS REPOSITORY IS NOW ARCHIVED IN FAVOR OF
> https://github.com/triggermesh/tmctl

The `tmctl` repository was archived on 2023-12-11 and the `README`
wasn't updated to explain the project status before archiving.

This presumably means that the project isn't being
developed/maintained further, so this deprecates the formula
accordingly.