gist 6.0.0 (undeprecate) - Revert "#161777"
This reverts commit 27d809277ed8102861524f66d163c5838e8de873, reversing
changes made to a7bebfef9231485c2340236c3dd6035761c9c868.

Reverts Homebrew/homebrew-core#161777

The PR that deprecated the gist formula was opened on Feb 4, 2024.
It was during the time that @defunkt 's account was (for who knows what reason) temporarily suspended.

He tweeted that his account was suspended on Feb 5:
https://twitter.com/defunkt/status/1754610843361362360
and access was restored a few hours later:
https://twitter.com/defunkt/status/1754663592186851610.

Presumably, the gist repo was inaccessible during the time that the account was suspended, and thus appeared as if the repo was gone.
It is not gone, and is still online and available: https://github.com/defunkt/gist