libxc: dont compile third/fourth derivative code
A bigger software will use libxc in the future, but with this functions. The development of this software is on a macos-system.
As i see inside the ChangeLog.md

"- Added the possibility of selectively disabling the compilation of vxc, fxc, kxc, and lxc to both Autotools and CMake build systems. By default kxc and lxc are not compiled. This allows for faster compilation times."

this parameter is inside just for reducing the compiling-time.

related upstream commits:
- https://gitlab.com/libxc/libxc/-/commit/b0ce1d8eaf224c733a00f1cbd02e85d7b85bc5ff
- https://gitlab.com/libxc/libxc/-/commit/a161a63499f882b9256815c218a1f007b0a055b7

Signed-off-by: Rui Chen <rui@chenrui.dev>