ehco: update livecheck
The existing `livecheck` block for `ehco` uses the `GithubLatest`
strategy as a way of working around some problematic tags. However,
the check is currently failing because the "latest" release on GitHub
is an unstable version (`v0.0.0-nightly`).

The problematic tags have been removed at this point, so this fixes
the check by switching back to the `Git` strategy with the standard
regex for tags like `1.2.3`/`v1.2.3` (which will omit the
aforementioned nightly tag).