ksync: add livecheck
By default, livecheck checks the Git tags for `ksync` and this works
but it continually returns `0.4.7-hotfix` as a newer version because
the formula version is 0.4.7 (parsed from the same tag).

This resolves the issue by adding a `livecheck` block that uses a
regex to restrict matching to tags like `1.2.3`. When upstream has
created hotfix releases, there has always been a previous tag for the
same version without the `-hotfix` suffix, so this should work fine
if that continues to be true.