workflows/tests: fix `tap_syntax` style cache key
Sometimes the `tap_syntax` job for `brew`'s `master` branch incorrectly
uses the cache from the `stable` job, because the cache restore key for
the former is a prefix of the key for the latter. This can be seen in
the following job, for example:
https://github.com/Homebrew/homebrew-core/actions/runs/8652032061/job/23724090864.

Fix that by adding `-master` to the `master` job's cache key, and
updating the restore key accordingly.