glasskube: fix missing web bundles
This was raised in issue https://github.com/Homebrew/homebrew-core/issues/168623.

The build process for Glasskube has changed, it now includes bundling of web resources.
If this step is skipped, some features don't work as expected.
To fix this problem, this commit changes the glasskube formula to include running "make web",
which is equivalent to something like "npm install && npm run build".

Signed-off-by: Jakob Steiner <jakob.steiner@glasskube.eu>