rake-completion: replace disabled formula
@JoeNyland's rake-completion was disabled (due to upstream repo being
archived) in January 2024. It has received 0 installations in 90 days.

This replacement is by @mernen who is the author of related completions
for gem, bundler, rails, and ruby (all of which are already in homebrew/core).

The two rake completion tools are _structurally_ compatible as they both
work with bash-completion. They are not cli tools or libraries so there
is no API to be compatible with.