workflows/tests: fix artifact name conflict in dep test logs
For dependent testing, we only add the `-deps` suffix to the name of
self-hosted runners (ephimeral self-hosted runners to be more precise),
so the runner name for Linux (e.g., `linux-self-hosted-1` or
`ubuntu-22.04`) is the same for both bottling and dep testing. This
resulted in an error [^1] while uploading logs for dep testing after
migration to artifact v4, as artifact names must be unique.

We can work around that by adding the `-deps` suffix for Linux runners
as well.

N.B. Although we test deps on macOS with GitHub-hosted runners too, we
don't do that for bottling so there should be no conflict as long as
this remains the case.

[^1]: https://github.com/Homebrew/homebrew-core/actions/runs/8742672877/job/23993144285