rtabmap: add livecheck
By default, livecheck checks the Git tags for `rtabmap` but it's
currently reporting `0.21.5-jazzy` as newest, though 0.21.4 is the
latest version with a GitHub release. Upstream doesn't create
releases for all of the tags and it seems that tags like
`0.21.5-jazzy` may not be stable versions, so this adds a `livecheck`
block that uses the `GithubLatest` strategy.