workflows/tests: enable more granular control of deps jobs
Currently, we request long timeouts for PRs whenever at least one of the
formula(e) tests or dependent tests require a long timeout.

This is wasteful of CI resources, because a typical PR that requires a
long timeout only requires it for only one of the formula(e) tests or
the dependent tests.

To avoid unnecessarily congesting the long build queue, let's allow more
granular control of which jobs get queued for a long build by using a
separate label for dependent testing jobs that require a long timeout.

I've also removed the hard-coded values for the short and long timeouts
from the `check-labels.js` script so that we can keep these in `brew`
instead (and now we'll only need to keep track of them in one place
instead of in two).