ansible: fix ssh issues with netconf
Recent updates to netconf (for working with routers) require pylibssh in order to
work consistently. This commit adds ansible-pylibssh in order to resolve errors when
communicating with routers using vyos and junos (at a minimum).

Adds libssh as a dependency, which is necessary for build and will cause an error
if run without it.