minikube v1.33.1 fix revision
minikube v1.33.1 was tagged and release automation was run on the tag.
The release automation failed due to an RPM build issue which would require merging a fix and retagging.
I deleted the tag and created a PR to resolve the RPM build issue.
The tag was only up for ~30 minutes, but Homebrew automation picked up the new tag during this window and created a PR to bump minikube.
After the PR was merged I retagged and triggered release automation.
Users are now reporting issues when trying to upgrade minikube in brew due to the commit of the tag not matching what is in brew.
Updating the revision to match the actual tag revision.