sevenzip: update livecheck, add head
The existing `livecheck` block for `sevenzip` checks the first-party
download page but it currently returns 23.01 as the latest version
instead of 24.06. The regex is only set up to match text like
"Download 7-Zip 23.01 (2023-06-20):" but the newest version uses
"Download 7-Zip 24.06 (2024-05-26) for Windows:". Each release
provides a source code tarball regardless of whether it's "for
Windows" or not, so this updates the regex to match all of the
releases.

This also adds a `head` URL, in case we need to check the GitHub
repository for version information in the future (using the
`GithubLatest` strategy). It seemed to build fine when testing with
`brew install --HEAD sevenzip`.