pdns: update livecheck
The existing `livecheck` block for `pdns` checks the directory
listing page where the `stable` archive is found. We usually only
check a directory listing page when we don't have a better option
(like a download page), as sometimes a new file may be uploaded
before it's officially released.

With that in mind, this updates the `livecheck` block to check the
first-party "Downloads" page instead. A nice side effect is that it
has a smaller download size (8.67 KB compressed vs. 142 KB, as the
directory listing server doesn't use compression) and a faster
response time (~150-300 ms vs. ~750-950 ms).